@@ -1,1 +1,1 @@
-export {run} from '@oclif/command'
+export { run } from '@oclif/command';
