@@ -2,4 +2,4 @@ export type Recipe = {
   tech: string;
   category: string;
   name: string;
-}
+};
